refactor: update register club activity-report section folder structure
