+++ packages/web/src/features/register-club/components/activity-report/CreateActivityReportModal.tsx
@@ -2,12 +2,11 @@ import { addHours } from "date-fns";
 import React, { useCallback } from "react";
 import { useForm } from "react-hook-form";
 
-import { ApiAct007RequestBody } from "@sparcs-clubs/interface/api/activity/endpoint/apiAct007";
-
 import Modal from "@sparcs-clubs/web/common/components/Modal";
+import { ActivityReportFormData } from "@sparcs-clubs/web/features/activity-report/types/form";
 import usePostActivityReportForNewClub from "@sparcs-clubs/web/features/register-club/services/usePostActivityReportForNewClub";
 
-import ActivityReportForm from "./ActivityReportForm";
+import ActivityReportForm from "./_atomic/ActivityReportForm";
 
 interface CreateActivityReportModalProps {
   clubId: number;
@@ -23,25 +22,27 @@ const CreateActivityReportModal: React.FC<CreateActivityReportModalProps> = ({
   close,
   refetch,
 }) => {
-  const formCtx = useForm<ApiAct007RequestBody>({ mode: "all" });
+  const formCtx = useForm<ActivityReportFormData>({ mode: "all" });
 
   const { mutate } = usePostActivityReportForNewClub();
 
   const submitHandler = useCallback(
-    (_data: ApiAct007RequestBody, e: React.BaseSyntheticEvent) => {
-      e.preventDefault();
+    (data: ActivityReportFormData) => {
       mutate(
         {
           body: {
-            ..._data,
+            ...data,
             clubId,
-            durations: _data.durations.map(({ startTerm, endTerm }) => ({
+            durations: data.durations.map(({ startTerm, endTerm }) => ({
               startTerm: addHours(startTerm, 9),
               endTerm: addHours(endTerm, 9),
             })),
-            evidence: _data.evidence ?? "", // NOTE: (@dora) evidence is optional
-            participants: _data.participants.map(({ studentId }) => ({
-              studentId,
+            evidence: data.evidence ?? "", // NOTE: (@dora) evidence is optional
+            evidenceFiles: data.evidenceFiles.map(({ id }) => ({
+              fileId: id,
+            })),
+            participants: data.participants.map(({ id }) => ({
+              studentId: id,
             })),
           },
         },
@@ -60,19 +61,12 @@ const CreateActivityReportModal: React.FC<CreateActivityReportModalProps> = ({
     close();
   };
 
-  const handleSubmit = (e: React.BaseSyntheticEvent) => {
-    formCtx.handleSubmit(_data => submitHandler(_data, e))();
-  };
-
   return (
     <Modal isOpen={isOpen} width="full">
       <ActivityReportForm
-        // TODO. 리팩토링 해야 함
-        /* eslint-disable  @typescript-eslint/no-explicit-any */
         clubId={clubId}
-        formCtx={formCtx as any}
         onCancel={handleCancel}
-        onSubmit={handleSubmit}
+        onSubmit={() => formCtx.handleSubmit(submitHandler)}
       />
     </Modal>
   );
